Move Live OS tear down task to Live OS source
