+++ pyanaconda/modules/payload/live/live_os.py
@@ -33,9 +33,9 @@ from pyanaconda.modules.payload.base.initialization import PrepareSystemForInsta
     CopyDriverDisksFilesTask
 from pyanaconda.modules.payload.base.utils import get_dir_size
 from pyanaconda.modules.payload.live.live_os_interface import LiveOSHandlerInterface
-from pyanaconda.modules.payload.live.initialization import TeardownInstallationSourceTask, \
-    UpdateBLSConfigurationTask
-from pyanaconda.modules.payload.sources.initialization import SetUpInstallationSourceTask
+from pyanaconda.modules.payload.live.initialization import UpdateBLSConfigurationTask
+from pyanaconda.modules.payload.sources.initialization import TearDownInstallationSourceTask, \
+    SetUpInstallationSourceTask
 from pyanaconda.modules.payload.live.installation import InstallFromImageTask
 from pyanaconda.modules.payload.live.utils import get_kernel_version_list
 
@@ -130,7 +130,7 @@ class LiveOSHandlerModule(PayloadHandlerBase):
 
     def teardown_installation_source_with_task(self):
         """Teardown installation source device."""
-        return TeardownInstallationSourceTask(INSTALL_TREE)
+        return TearDownInstallationSourceTask(INSTALL_TREE)
 
     def pre_install_with_task(self):
         """Prepare intallation task."""
